+++ src/main/kotlin/frc/team3324/library/commands/PIDCommand.kt
@@ -1,18 +1,21 @@
-package frc.team3324.library.commands
+package frc.team3324.libarry.commands
 
+import edu.wpi.first.wpilibj.Notifier
 import edu.wpi.first.wpilibj2.command.CommandBase
-import edu.wpi.first.wpilibj2.command.SubsystemBase
+import edu.wpi.first.wpilibj2.command.Subsystem
 
-class PIDCommand(val kP: Double, val kI: Double, val kD: Double, val goal: Double, val dt: Double, val subsystem: SubsystemBase, val measurement: () -> Double, val useOutput: (Double) -> Unit): CommandBase() {
+class PIDCommand(val kP: Double, val kI: Double, val kD: Double, val goal: Double, val dt: Double, subsystem: Subsystem, val measurement: () -> Double, val useOutput: (Double) -> Unit): CommandBase() {
     private var integral = 0.0
     private var lastPosition = 0.0
+    private val notifier = Notifier(this ::executePID)
 
     init {
-        this.addRequirements(subsystem)
+        addRequirements(subsystem)
     }
 
     override fun initialize() {
         integral = 0.0
+        notifier.startPeriodic(dt)
     }
 
     private fun executePID() {
@@ -27,7 +30,17 @@ class PIDCommand(val kP: Double, val kI: Double, val kD: Double, val goal: Doubl
         useOutput(output)
     }
 
-    override fun isFinished(): Boolean {
-        return measurement() == goal // return true if error is 0
+    override fun end(interrupted: Boolean) {
+        stopNotifier()
     }
+
+    fun stopNotifier() {
+        notifier.stop()
+        notifier.stop()
+    }
+
+    override fun isFinished(): Boolean {
+        return false
+    }
+
 }
