+++ src/main/kotlin/frc/team3324/library/commands/ToggleLightCommand.kt
@@ -1,10 +1,10 @@
-package frc.team3324.robot.util
+package frc.team3324.library.commands
 
 import edu.wpi.first.wpilibj.DigitalOutput
 import edu.wpi.first.wpilibj2.command.CommandBase
 import edu.wpi.first.wpilibj2.command.InstantCommand
 
-class ToggleLight(val light: DigitalOutput): InstantCommand() {
+class ToggleLightCommand(val light: DigitalOutput): InstantCommand() {
     override fun execute() {
         light.set(!light.get())
     }
